Calculate the value of the density function

Calculate the value of the density function, given specification as either parametric or Royston-Parmar splines model

Usage

calc_dens(time, type, spec)

Arguments

time

is the time at which the distribution function should be calculated.

type

is either "par" for regular parametric form (exponential, weibull etc) or "spl" for Royston-Parmar splines.

spec

is a list comprising: If type=="par": dist is the statistical distribution (named per flexsurv::flexsurvreg) and pars is a vector of the parameters for that distribution.

Exponential distribution (exp) requires the rate parameter.

Weibull distribution (both weibullPH and weibull formulations) requires the shape and scale parameters.

Log-logistic distribution (llogis) requires the shape and scale parameters.

Log-normal distribution (lnorm) requires the meanlog and sdlog parameters.

Gamma and Gompertz distributions (gamma and gompertz) require the shape and rate parameters.

Generalized Gamma requires the mu, sigma and Q parameters if using the standard parameterization (gengamma) or shape, scale and k parameters if using the original parameterization (gengamma.orig). If type=="spl":

gamma - Vector of parameters describing the baseline spline function, as described in flexsurv::flexsurvspline. This may be supplied as a vector with number of elements equal to the length of knots, in which case the parameters are common to all times. Alternatively a matrix may be supplied, with rows corresponding to different times, and columns corresponding to knots.

knots - Vector of locations of knots on the axis of log time, supplied in increasing order. Unlike in flexsurv::flexsurvspline, these include the two boundary knots.

scale - Either "hazard", "odds", or "normal", as described in flexsurv::flexsurvspline. With the default of no knots in addition to the boundaries, this model reduces to the Weibull, log-logistic and log-normal respectively. The scale must be common to all times.

Value

The value of a hazard function, a numeric value.

Examples

calc_dens(time=1:5,
    type="spl",
    spec=list(gamma=c(0.1,0.2,0.1), knots=c(-5,2,4), scale="normal")
    )
#> [1] 0 0 0 0 0
calc_dens(time=1:5,
    type="par",
    spec=list(dist="lnorm", pars=c(3,1))
    )
#> [1] 0.004431848 0.013941703 0.021814325 0.027126183 0.030342439