Calculate likelihoods for three three-state model structures

Calculate likelihood values and other summary output for the following three state models structures: partitioned survival, clock forward state transition, and clock reset state transition. The function requires appropriately formatted patient-level data, a set of fitted survival regressions, and the time cut-off (if two-piece modeling is used).

Usage

calc_likes(ptdata, dpam, cuttime = 0)

Arguments

ptdata

Dataset of patient level data. Must be a tibble with columns named:

ptid: patient identifier

pfs.durn: duration of PFS from baseline

pfs.flag: event flag for PFS (=1 if progression or death occurred, 0 for censoring)

os.durn: duration of OS from baseline

os.flag: event flag for OS (=1 if death occurred, 0 for censoring)

ttp.durn: duration of TTP from baseline (usually should be equal to pfs.durn)

ttp.flag: event flag for TTP (=1 if progression occurred, 0 for censoring).

Survival data for all other endpoints (time to progression, pre-progression death, post-progression survival) are derived from PFS and OS.

dpam

List of survival regressions for each endpoint:

pre-progression death (PPD)

time to progression (TTP)

progression-free survival (PFS)

overall survival (OS)

post-progression survival clock forward (PPS-CF) and

post-progression survival clock reset (PPS-CR).

cuttime

Time cutoff - this is nonzero for two-piece models.

Value

Two outputs are returned:

results is a tibble of values and data relating to the likelihood for this model:

npts: Number of patients analysed for each endpoint.

likedata: Patient-level dataset with additional likelihood-related calculations.

coefsdists: Summary table of distributions and parameters used for each endpoint.

slikes: Total log-likelihood for each possible outcome

ll: Total log-likelihood

params: Number of parameters used in this model

AIC: Akaike Information Criterion value for this model

BIC: Bayesian Information Criterion value for this model

llcomp is a tibble providing a breakdown of the likelihood calculations by outcome. Outcomes are as follows:

(1) refers to patients who remain alive and progression-free during the follow-up;

(2) refers to patients who die without prior progression during the follow-up;

(3) refers to patients who progress and then remain alive for the remaining follow-up, and

(4) refers to patients who progress and die within the follow-up.

The number of patients for each outcome are given for each model structure. You may confirm that these are identical across model structures. The contribution of each patient group to the calculation of log-likelihood for each model is given in fields beginning ll_. This is helpful for understanding differences in likelihoods between model structures, according to patient outcomes.

See also

Examples

bosonc <- create_dummydata("flexbosms")
fits <- fit_ends_mods_spl(bosonc)
# Pick out best distribution according to min AIC
params <- list(
  ppd = find_bestfit_spl(fits$ppd, "aic")$fit,
  ttp = find_bestfit_spl(fits$ttp, "aic")$fit,
  pfs = find_bestfit_spl(fits$pfs, "aic")$fit,
  os = find_bestfit_spl(fits$os, "aic")$fit,
  pps_cf = find_bestfit_spl(fits$pps_cf, "aic")$fit,
  pps_cr = find_bestfit_spl(fits$pps_cr, "aic")$fit
  )
calc_likes(bosonc, dpam=params)
#> $detailed
#> # A tibble: 6 × 7
#>   valid outcome  npts   ll_1   ll_2   ll_3   ll_4
#>   <lgl>   <dbl> <int>  <dbl>  <dbl>  <dbl>  <dbl>
#> 1 FALSE       4     1   NA     NA     NA     NA  
#> 2 TRUE        1    72  -64.8  -64.8  -65.2  -65.2
#> 3 TRUE        2    29 -152.  -147.  -147.  -147. 
#> 4 TRUE        3    35 -151.  -150.  -148.  -148. 
#> 5 TRUE        4    67 -469.    NA     NA     NA  
#> 6 TRUE        4    68   NA   -473.  -473.  -470. 
#> 
#> $valid
#> # A tibble: 4 × 6
#>   methno methname    npts_TRUE npts_FALSE ll_TRUE ll_FALSE
#>    <dbl> <chr>           <int>      <int>   <dbl>    <dbl>
#> 1      1 psm_simple        203          1   -836.    NA   
#> 2      2 psm_complex       203          1   -830.    -5.25
#> 3      3 stm_cf            203          1   -828.    -5.45
#> 4      4 stm_cr            203          1   -824.    -6.08
#> 
#> $sumall
#> # A tibble: 8 × 10
#>   validall methno methname     npts      ll nparam   aic   bic rank_aic rank_bic
#>   <lgl>     <dbl> <chr>       <int>   <dbl>  <dbl> <dbl> <dbl>    <dbl>    <dbl>
#> 1 FALSE         1 psm_simple      1   NA         7   NA    NA        NA       NA
#> 2 FALSE         2 psm_complex     1   -5.25      9   NA    NA        NA       NA
#> 3 FALSE         3 stm_cf          1   -5.45      9   NA    NA        NA       NA
#> 4 FALSE         4 stm_cr          1   -6.08     10   NA    NA        NA       NA
#> 5 TRUE          1 psm_simple    203 -836.        7 1685. 1709.        4        4
#> 6 TRUE          2 psm_complex   203 -830.        9 1679. 1708.        3        3
#> 7 TRUE          3 stm_cf        203 -828.        9 1673. 1703.        2        2
#> 8 TRUE          4 stm_cr        203 -824.       10 1668. 1701.        1        1
#>