Calculate likelihood for a three-state clock forward state transition model

Calculate likelihood values and other summary output for a three-state clock forward state transition model, given appropriately formatted patient-level data, a set of fitted survival regressions, and the time cut-off (if two-piece modeling is used). This function is called by calc_likes.

Usage

calc_likes_stm_cf(ptdata, dpam, cuttime = 0)

Arguments

ptdata

Dataset of patient level data. Must be a tibble with columns named:

ptid: patient identifier

pfs.durn: duration of PFS from baseline

pfs.flag: event flag for PFS (=1 if progression or death occurred, 0 for censoring)

os.durn: duration of OS from baseline

os.flag: event flag for OS (=1 if death occurred, 0 for censoring)

ttp.durn: duration of TTP from baseline (usually should be equal to pfs.durn)

ttp.flag: event flag for TTP (=1 if progression occurred, 0 for censoring).

Survival data for all other endpoints (time to progression, pre-progression death, post-progression survival) are derived from PFS and OS.

dpam

List of survival regressions for each endpoint:

pre-progression death (PPD)

time to progression (TTP)

progression-free survival (PFS)

overall survival (OS)

post-progression survival clock forward (PPS-CF) and

post-progression survival clock reset (PPS-CR).

cuttime

Time cutoff - this is nonzero for two-piece models.

Value

List of values and data relating to the likelihood for this model:

npts: Number of patients analysed for each endpoint.

likedata: Patient-level dataset with additional likelihood-related calculations.

coefsdists: Summary table of distributions and parameters used for each endpoint.

slikes: Total log-likelihood for each possible outcome

ll: Total log-likelihood

params: Number of parameters used in this model

AIC: Akaike Information Criterion value for this model

BIC: Bayesian Information Criterion value for this model

See also

Examples

bosonc <- create_dummydata("flexbosms")
fits <- fit_ends_mods_spl(bosonc)
# Pick out best distribution according to min AIC
params <- list(
  ppd = find_bestfit_spl(fits$ppd, "aic")$fit,
  ttp = find_bestfit_spl(fits$ttp, "aic")$fit,
  pfs = find_bestfit_spl(fits$pfs, "aic")$fit,
  os = find_bestfit_spl(fits$os, "aic")$fit,
  pps_cf = find_bestfit_spl(fits$pps_cf, "aic")$fit,
  pps_cr = find_bestfit_spl(fits$pps_cr, "aic")$fit
  )
calc_likes_stm_cf(bosonc, dpam=params)
#> $npts
#> [1] 204 204
#> 
#> $npar
#> [1] 9
#> 
#> $ll
#> [1] -833.162 -833.162
#> 
#> $data
#> # A tibble: 204 × 23
#>     ptid pfs.durn pfs.flag os.durn os.flag ttp.durn ttp.flag  httpu sttpu
#>    <int>    <dbl>    <dbl>   <dbl>   <dbl>    <dbl>    <dbl>  <dbl> <dbl>
#>  1     1   18.7          1   42.9        1   18.7          1 0.0326 0.448
#>  2     2   12.0          1   23.3        1   12.0          1 0.0387 0.567
#>  3     3    0.452        1    8.81       1    0.452        1 0.0623 0.971
#>  4     4    9.07         1   52.7        1    9.07         1 0.0421 0.638
#>  5     5   47.7          0   47.7        0   47.7          0 0.0196 0.218
#>  6     6    3.26         1   13.1        1    3.26         1 0.0502 0.833
#>  7     7    0.738        1    2.74       1    0.738        1 0.0594 0.954
#>  8     8   35.5          1   35.5        1   35.5          0 0.0233 0.282
#>  9     9   10.4          1   37.1        1   10.4          1 0.0406 0.605
#> 10    10   10.5          1   32.7        1   10.5          1 0.0405 0.603
#> # ℹ 194 more rows
#> # ℹ 14 more variables: hppdu <dbl>, sppdu <dbl>, sppsu <dbl>, sppst <dbl>,
#> #   hppst <dbl>, f1 <dbl>, f2 <dbl>, f3 <dbl>, f4 <dbl>, like <dbl>,
#> #   llike <dbl>, outcome <dbl>, chf <lgl>, valid <lgl>
#>