Find the "best" survival regression from a list

When there are multiple survival regressions fitted to the same endpoint and dataset, it is necessary to identify the preferred model. This function reviews the fitted regressions and selects that with the minimum Akaike or Bayesian Information Criterion (AIC, BIC), depending on user choice.

Usage

find_bestfit_par(reglist, crit)

Arguments

reglist

List of fitted survival regressions to an endpoint and dataset.

crit

Criterion to be used in selection of best fit, either "aic" (Akaike Information Criterion) or "bic" (Bayesian Information Criterion).

Value

List of the single survival regression with the best fit.

Examples

bosonc <- create_dummydata("flexbosms")
# Pick some distributions to fit to all endpoints
dists <- c("exp", "llogis", "lnorm")
# Fit all distributions to all endpoints
fits <- fit_ends_mods_par(bosonc,
    cuttime = 0,
    ppd.dist = dists,
    ttp.dist = dists,
    pfs.dist = dists,
    os.dist = dists,
    pps_cr.dist = dists,
    pps_cf.dist = dists)
find_bestfit_par(fits$ttp, "aic")
#> $fit
#> Call:
#> .f(formula = ..1, dist = ..2)
#> 
#> Estimates: 
#>        est      L95%     U95%     se     
#> shape   1.0300   0.8726   1.2158   0.0872
#> scale  15.4175  11.8479  20.0625   2.0716
#> 
#> N = 204,  Events: 103,  Censored: 101
#> Total time at risk: 2358.845
#> Log-likelihood = -422.068, df = 2
#> AIC = 848.136
#> 
#> 
#> $results
#> # A tibble: 3 × 11
#>      id dists   npts  pars loglik conv  posdef   aic   bic rankaic rankbic
#>   <int> <chr>  <dbl> <dbl>  <dbl> <lgl> <lgl>  <dbl> <dbl>   <dbl>   <dbl>
#> 1     1 exp      204     1  -426. TRUE  TRUE    853.  856.       3       2
#> 2     2 llogis   204     2  -422. TRUE  TRUE    848.  855.       1       1
#> 3     3 lnorm    204     2  -424. TRUE  TRUE    851.  858.       2       3
#> 
find_bestfit_par(fits$os, "bic")
#> $fit
#> Call:
#> .f(formula = ..1, dist = ..2)
#> 
#> Estimates: 
#>        est     L95%    U95%    se    
#> shape   1.814   1.541   2.136   0.151
#> scale  23.832  20.361  27.895   1.914
#> 
#> N = 204,  Events: 97,  Censored: 107
#> Total time at risk: 3479.143
#> Log-likelihood = -436.9493, df = 2
#> AIC = 877.8987
#> 
#> 
#> $results
#> # A tibble: 3 × 11
#>      id dists   npts  pars loglik conv  posdef   aic   bic rankaic rankbic
#>   <int> <chr>  <dbl> <dbl>  <dbl> <lgl> <lgl>  <dbl> <dbl>   <dbl>   <dbl>
#> 1     1 exp      204     1  -444. TRUE  TRUE    890.  894.       3       3
#> 2     2 llogis   204     2  -437. TRUE  TRUE    878.  885.       1       1
#> 3     3 lnorm    204     2  -439. TRUE  TRUE    883.  889.       2       2
#>