Fit multiple parametric survival regressions to the multiple required endpoints

Fits multiple parametric survival regressions, according to the distributions stipulated, to the multiple endpoints required in fitting partitioned survival analysis, clock forward and clock reset semi-markov models.

Usage

fit_ends_mods_par(
  ds,
  cuttime = 0,
  ppd.dist = c("exp", "weibullPH", "llogis", "lnorm", "gamma", "gompertz"),
  ttp.dist = c("exp", "weibullPH", "llogis", "lnorm", "gamma", "gompertz"),
  pfs.dist = c("exp", "weibullPH", "llogis", "lnorm", "gamma", "gompertz"),
  os.dist = c("exp", "weibullPH", "llogis", "lnorm", "gamma", "gompertz"),
  pps_cf.dist = c("exp", "weibullPH", "llogis", "lnorm", "gamma", "gompertz"),
  pps_cr.dist = c("exp", "weibullPH", "llogis", "lnorm", "gamma", "gompertz"),
  expvar = NA
)

Arguments

ds

Dataset of patient level data. Must be a tibble with columns named:

ptid: patient identifier

pfs.durn: duration of PFS from baseline

pfs.flag: event flag for PFS (=1 if progression or death occurred, 0 for censoring)

os.durn: duration of OS from baseline

os.flag: event flag for OS (=1 if death occurred, 0 for censoring)

ttp.durn: duration of TTP from baseline (usually should be equal to pfs.durn)

ttp.flag: event flag for TTP (=1 if progression occurred, 0 for censoring).

Survival data for all other endpoints (time to progression, pre-progression death, post-progression survival) are derived from PFS and OS.

cuttime

Cut-off time for a two-piece model, equals zero for one-piece models.

ppd.dist

Vector of distributions (named per flexsurv::flexsurvreg()) to be fitted to Pre-Progression Death (PPD).

ttp.dist

Vector of distributions (named per flexsurv::flexsurvreg()) to be fitted to Time To Progression (TTP).

pfs.dist

Vector of distributions (named per flexsurv::flexsurvreg()) to be fitted to Progression-Free Survival (PFS).

os.dist

Vector of distributions (named per flexsurv::flexsurvreg()) to be fitted to Overall Survival (OS).

pps_cf.dist

Vector of distributions (named per flexsurv::flexsurvreg()) to be fitted to Post Progression Survival, where time is from baseline (clock forward).

pps_cr.dist

Vector of distributions (named per flexsurv::flexsurvreg()) to be fitted to Post Progression Survival, where time is from progression (clock reset).

expvar

Explanatory variable for modeling of PPS

Value

A list by endpoint, then distribution, each containing two components:

result: A list of class flexsurvreg containing information about the fitted model.

error: Any error message returned on fitting the regression (NULL indicates no error). Also, the given cuttime.

See also

Examples

bosonc <- create_dummydata("flexbosms")
fit_ends_mods_par(bosonc, expvar=bosonc$ttp.durn)
#> Error in fit_ends_mods_par(bosonc, expvar = bosonc$ttp.durn): object 'pps.durn' not found