Fit multiple spline regressions to the multiple required endpoints

Fits multiple survival regressions, according to the distributions stipulated, to the multiple endpoints required in fitting partitioned survival analysis, clock forward and clock reset semi-markov models.

Usage

fit_ends_mods_spl(
  ds,
  knot_set = 1:3,
  scale_set = c("hazard", "odds", "normal"),
  expvar = NA
)

Arguments

ds

Dataset of patient level data. Must be a tibble with columns named:

ptid: patient identifier

pfs.durn: duration of PFS from baseline

pfs.flag: event flag for PFS (=1 if progression or death occurred, 0 for censoring)

os.durn: duration of OS from baseline

os.flag: event flag for OS (=1 if death occurred, 0 for censoring)

ttp.durn: duration of TTP from baseline (usually should be equal to pfs.durn)

ttp.flag: event flag for TTP (=1 if progression occurred, 0 for censoring).

Survival data for all other endpoints (time to progression, pre-progression death, post-progression survival) are derived from PFS and OS.

knot_set

A vector of the number of knots to fit spline models to. For example, knot_set=1:3 will result in spline models fitted to 1, 2 and 3 (internal) knot splines.

scale_set

A vector of the scales of knots to fit spline models to. For example, scale_set=c("hazard", "odds", "normal") will result in spline models fitted on hazard, odds and normal scales.

expvar

Explanatory variable for modeling of PPS

Value

A list by endpoint, then distribution, each containing two components:

result: A list of class flexsurv::flexsurvspline containing information about the fitted model.

error: Any error message returned on fitting the regression (NULL indicates no error). Also, the given cuttime.

See also

Examples

# Create dataset in suitable form using bos dataset from the flexsurv package
bosonc <- create_dummydata("flexbosms")
fit_ends_mods_spl(bosonc, expvar=bosonc$ttp.durn)
#> Error in fit_ends_mods_spl(bosonc, expvar = bosonc$ttp.durn): object 'pps.durn' not found