Fit survival regressions with multiple parametric distributions

Fits survival regressions with flexsurv::flexsurvreg to multiple statistical distributions in one function call.

Usage

fit_mods_par(durn1, durn2 = NA, evflag, dists, expvar = NA)

Arguments

durn1

First time point, corresponds with time in survival::Surv(). For right censored data, this is the follow up time. For interval data, the first argument is the starting time for the interval.

durn2

Second time point, corresponds with time2 in survival::Surv(). The ending time of the interval for interval censored or counting process data only.

evflag

Event flag, corresponds with event in survival::Surv. The status indicator is normally 0=alive, 1=dead.

dists

is a vector of all the distributions (named according to flexsurv::flexsurvreg()) for which statistical fits are requested.

expvar

Explanatory variable for modeling of PPS

Value

A list by distribution, each containing two components:

result: A list of class flexsurv::flexsurvreg containing information about the fitted model.

error: Any error message returned on fitting the regression (NULL indicates no error).